requirements.txt
cffi==1.15.0
cityscapesscripts==2.2.0
click==8.0.4
Cython==0.29.32
detectron2==0.6+cu113
easydict==1.9
fvcore==0.1.5.post20220305
h5py==3.7.0
iopath==0.1.9
ipython==8.4.0
lxml==4.9.1
matplotlib==3.5.1
numpy==1.22.0
Pillow==9.2.0
pyparsing==3.0.7
PyYAML==6.0
scipy==1.7.3
setuptools==58.0.4
tensorboard==2.8.0
tensorboardX==2.5.1
tensorflow==2.9.1
torch==1.10.1
torchvision==0.11.2
tqdm==4.63.0
yacs==0.1.8
pycocotools==2.0.4
Werkzeug==2.0.3
